NewGridElement: UseCase redone fewer int-casts

+++ trunk/Baselet/src/com/baselet/control/Utils.java
@@ -83,7 +83,7 @@ public abstract class Utils {
 		return returnVector;
 	}
 
-	public static List<String> splitString(String text, int width) {
+	public static List<String> splitString(String text, double width) {
 		DiagramHandler handler = Main.getInstance().getDiagramHandler();
 		StringBuilder stringBuilder = new StringBuilder(text);
 		int lastEmptyChar = -1; // is -1 if there was no ' ' in this line
